// frontend/features/tasks/types/task.ts
export type TaskStatus =
  | "todo"
  | "in_progress"
  | "completed";

export type TaskPriority =
  | "low"
  | "medium"
  | "high";

export interface Task {
  id: string;
  title: string;
  description: string;

  status: TaskStatus;
  priority: TaskPriority;

  dueDate: string | null;

  goalId: string | null;

  createdAt: string;
  updatedAt: string;
}